+++ ros/src/computing/perception/localization/packages/lidar_localizer/nodes/ndt_matching_monitor/ndt_matching_monitor_node.cpp
@@ -40,7 +40,7 @@ int main(int argc, char **argv)
 {
 	ros::init(argc, argv, __APP_NAME__);
 
-	RosNdtMatchingMonitor app;
+	RosNDTMatchingMonitor app;
 
 	app.Run();