+++ ros/src/util/packages/data_preprocessor/scripts/rtmgr.py
@@ -773,9 +773,9 @@ class MyDialogLaneStop(wx.Dialog):
 
 # end of class MyDialogLaneStop
 
-class MyDialogNdtMapping(wx.Dialog):
+class MyDialogNDTMapping(wx.Dialog):
 	def __init__(self, *args, **kwds):
-		# begin wxGlade: MyDialogNdtMapping.__init__
+		# begin wxGlade: MyDialogNDTMapping.__init__
 		kwds["style"] = wx.CAPTION
 		wx.Dialog.__init__(self, *args, **kwds)
 		self.panel_v = wx.ScrolledWindow(self, wx.ID_ANY, style=wx.TAB_TRAVERSAL)
@@ -800,7 +800,7 @@ class MyDialogNdtMapping(wx.Dialog):
 		# end wxGlade
 
 	def __set_properties(self):
-		# begin wxGlade: MyDialogNdtMapping.__set_properties
+		# begin wxGlade: MyDialogNDTMapping.__set_properties
 		self.SetTitle(_("ndt_mapping"))
 		self.SetSize((352, 341))
 		self.panel_v.SetScrollRate(10, 10)
@@ -808,7 +808,7 @@ class MyDialogNdtMapping(wx.Dialog):
 		# end wxGlade
 
 	def __do_layout(self):
-		# begin wxGlade: MyDialogNdtMapping.__do_layout
+		# begin wxGlade: MyDialogNDTMapping.__do_layout
 		sizer_30 = wx.BoxSizer(wx.VERTICAL)
 		sizer_48 = wx.BoxSizer(wx.VERTICAL)
 		sizer_49 = wx.BoxSizer(wx.HORIZONTAL)
@@ -830,23 +830,23 @@ class MyDialogNdtMapping(wx.Dialog):
 		self.Layout()
 		# end wxGlade
 
-	def OnRef(self, event):  # wxGlade: MyDialogNdtMapping.<event_handler>
+	def OnRef(self, event):  # wxGlade: MyDialogNDTMapping.<event_handler>
 		print "Event handler 'OnRef' not implemented!"
 		event.Skip()
 
-	def OnRadio(self, event):  # wxGlade: MyDialogNdtMapping.<event_handler>
+	def OnRadio(self, event):  # wxGlade: MyDialogNDTMapping.<event_handler>
 		print "Event handler 'OnRadio' not implemented!"
 		event.Skip()
 
-	def OnPcdOutput(self, event):  # wxGlade: MyDialogNdtMapping.<event_handler>
+	def OnPcdOutput(self, event):  # wxGlade: MyDialogNDTMapping.<event_handler>
 		print "Event handler 'OnPcdOutput' not implemented!"
 		event.Skip()
 
-	def OnOk(self, event):  # wxGlade: MyDialogNdtMapping.<event_handler>
+	def OnOk(self, event):  # wxGlade: MyDialogNDTMapping.<event_handler>
 		print "Event handler 'OnOk' not implemented!"
 		event.Skip()
 
-# end of class MyDialogNdtMapping
+# end of class MyDialogNDTMapping
 
 class MyDialogDpm(wx.Dialog):
 	def __init__(self, *args, **kwds):